Add ShoppingCart service implementation

+++ src/main/java/yablunovskyi/bookstore/dto/shoppingcart/UpdateCartItemRequestDto.java
@@ -6,5 +6,5 @@ import jakarta.validation.constraints.Positive;
 public record UpdateCartItemRequestDto(
         @NotNull
         @Positive
-        Long quantity) {
+        Integer quantity) {
 }
